+++ src/pythonwrench/csv.py
@@ -2,3 +2,11 @@
 # -*- coding: utf-8 -*-
 
 # for backward compatibility
+from pythonwrench.serialization.csv import (  # noqa: F401
+    dump_csv,
+    dumps_csv,
+    load_csv,
+    loads_csv,
+    read_csv,
+    save_csv,
+)