+++ src/pythonwrench/json.py
@@ -2,3 +2,11 @@
 # -*- coding: utf-8 -*-
 
 # for backward compatibility
+from pythonwrench.serialization.json import (  # noqa: F401
+    dump_json,
+    dumps_json,
+    load_json,
+    loads_json,
+    read_json,
+    save_json,
+)