+++ src/pythonwrench/jsonl.py
@@ -2,3 +2,11 @@
 # -*- coding: utf-8 -*-
 
 # for backward compatibility
+from pythonwrench.serialization.jsonl import (  # noqa: F401
+    dump_jsonl,
+    dumps_jsonl,
+    load_jsonl,
+    loads_jsonl,
+    read_jsonl,
+    save_jsonl,
+)