+++ src/pythonwrench/pickle.py
@@ -2,3 +2,11 @@
 # -*- coding: utf-8 -*-
 
 # for backward compatibility
+from pythonwrench.serialization.pickle import (  # noqa: F401
+    dump_pickle,
+    dumps_pickle,
+    load_pickle,
+    loads_pickle,
+    read_pickle,
+    save_pickle,
+)