+++ tests/test_disk_cache.py
@@ -3,6 +3,7 @@
 
 import random
 import unittest
+from typing import Dict, List
 from unittest import TestCase
 
 import pythonwrench as pw
@@ -37,6 +38,23 @@ class TestDiskCache(TestCase):
 
         assert data1 == data2
         assert data1 != data3
+
+    def test_disk_cache_example_3(self) -> None:
+        cache_fname = "disk_cache_example_3.csv"
+
+        @pw.disk_cache_decorator(
+            cache_saving_backend="csv",
+            cache_fname_fmt=cache_fname,
+        )
+        def disk_cache_example_3(num: int) -> List[Dict[str, str]]:
+            return pw.dict_list_to_list_dict(
+                {"a": ["a"] * num, "b": ["b"] * num}, "same"
+            )
+
+        outputs = disk_cache_example_3(10)
+
+        assert disk_cache_example_3.fn(10) == outputs  # type: ignore
+        assert (disk_cache_example_3.cache_fn_dpath / cache_fname).is_file()  # type: ignore
 
 
 if __name__ == "__main__":