+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## [0.5.3] UNRELEASED
 ### Fixed
 - `Ticker` class in global context.
+- `is_special_form` and `is_parametrized` now supports objects with custom `__contains__`, like ndarray or DataFrames.
 
 ## [0.5.2] 2026-07-02
 ### Modified